pyproject.toml
[project]
name = "front-page-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "feedparser>=6.0.11",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.3",
]
